+++ python/GafferSceneTest/__init__.py
@@ -171,6 +171,7 @@ from .MeshSegmentsTest import MeshSegmentsTest
 from .ImageToPointsTest import ImageToPointsTest
 from .MeshSplitTest import MeshSplitTest
 from .FramingConstraintTest import FramingConstraintTest
+from .MeshNormalsTest import MeshNormalsTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *